cleanup
git-svn-id: https://only.mawhrin.net/repos/FBReaderJ/trunk@378 6a642e6f-84f6-412e-ac94-c4a38d5a04b0

+++ src/org/zlibrary/text/view/ZLTextStyle.java
@@ -14,5 +14,5 @@ public interface ZLTextStyle {
 	public int spaceBefore();
 	public int spaceAfter();
 	public boolean isDecorated();
-	public int alignment();
+	public byte alignment();
 }
