+++ src/org/zlibrary/text/view/style/ZLTextFullStyleDecoration.java
@@ -19,7 +19,7 @@ public class ZLTextFullStyleDecoration extends ZLTextStyleDecoration {
 
 	public ZLDoubleOption LineSpaceOption;	
 
-	public ZLTextFullStyleDecoration(String name, int fontSizeDelta, ZLBoolean3 bold, ZLBoolean3 italic, int spaceBefore, int spaceAfter, int leftIndent,int rightIndent, int firstLineIndentDelta, int verticalShift, int alignment, double lineSpace, ZLBoolean3 allowHyphenations) {
+	public ZLTextFullStyleDecoration(String name, int fontSizeDelta, ZLBoolean3 bold, ZLBoolean3 italic, int spaceBefore, int spaceAfter, int leftIndent,int rightIndent, int firstLineIndentDelta, int verticalShift, byte alignment, double lineSpace, ZLBoolean3 allowHyphenations) {
 		super(name, fontSizeDelta, bold, italic, verticalShift, allowHyphenations);
 //		System.out.println("Constructed, fonSizeDelta = " + fontSizeDelta);
 		SpaceBeforeOption = null;
