+++ src/org/zlibrary/text/view/style/ZLTextStyleCollection.java
@@ -131,7 +131,7 @@ public class ZLTextStyleCollection {
 						int rightIndent = intValue(attributes, "rightIndent");
 						int firstLineIndentDelta = intValue(attributes, "firstLineIndentDelta");
 
-						int alignment = ZLTextAlignmentType.ALIGN_UNDEFINED;
+						byte alignment = ZLTextAlignmentType.ALIGN_UNDEFINED;
 						String alignmentString = attributes.get("alignment");
 						if (alignmentString != null) {
 							if (alignmentString.equals("left")) {
